[FIX] account: suggest default repartition lines
`default_get` is bad m'kay?

Reproduce:
* Accounting > Configuration > Taxes > Create
* Just add a name, save.

There are no repartition lines, but there should be at least one line
for the base line.
This happens because `default_get` was only computing the main field
`repartition_line_ids` and it was allowed because the contraint was not
checking that same field.

closes odoo/odoo#129840

X-original-commit: c964c6a5cfc461d764b13246ea54fb9437c202b7
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>